requirements.txt
line-bot-sdk
flask==2.2.3
python-dotenv==1.0.0
ngrok==0.8.1
boto3==1.21.59
